Breaking up manifest.go
Splitting significant types out to their own files
DCOPS-7525
